+++ libft/libft.h
@@ -20,14 +20,14 @@
 # include <grp.h>
 # include <time.h>
 
-int     ft_char(char *s, char c);
+int		ft_char(char *s, char c);
 char	*ft_strdup(const char *s1);
-int     ft_strcmp(const char *s1, const char *s2);
+int		ft_strcmp(const char *s1, const char *s2);
 void	ft_putstr(char const *s);
 void	ft_putchar(char c);
 char	*ft_strsub(char const *s, unsigned int start, size_t len);
-void	ft_putnbr(int n);
+void	ft_putnbr(long long n);
 char	*ft_strjoin(char const *s1, char const *s2);
-int	ft_strlen(const char *s);
+int		ft_strlen(const char *s);
 
 #endif
